feat: make download mode always available
Currently, SteamOS mode is only enabled when a download mode limit is
set. This is due to it being designed with the Steam Deck in mind and
lowering the TDP. For the Steam Deck, let us accept it even though it
exhibits the same problems as below.

On normal hardware with modern stanby states (s0ix), we have the following:
 - Steam consumes a lot of power when downloading games
 - This power exceeds the power limits of the s0ix sleep state as
   defined by Windows in which e.g., the Legion Go S turns off its fan.
 - If we limit the power envelope, downloads will take a lot longer
 - Therefore, download mode can only work in the "Screen Off" s0ix state,
   which has a normal TDP draw.

As such, setting a TDP limit for download mode for other hardware is
pointless. So, stub out the call when no limit is set.

Besides, other hardware has a functional suspend light/is designed for
these types of wakeups, so we can just let it run at full TDP.